some refactoring to reduce unreasonble implementation . (#242)
* some refactoring to reduce unreasonble implementation .

* comment update .
