+++ MissionEditor/MissionEditor.vcxproj.filters
@@ -303,6 +303,9 @@
     <ClCompile Include="TriggerDatabase.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="FileOpenDlg.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="res\VersionInfo.rc">
@@ -641,6 +644,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="TriggerDatabase.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="FileOpenDlg.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
