+++ MissionEditor/functions.cpp
@@ -1525,126 +1525,3 @@ CComPtr<IDirectDrawSurface7> BitmapToSurface(IDirectDraw7* pDD, const CBitmap& b
 
 	return pSurface;
 }
-
-CFileDialogClsid::CFileDialogClsid(
-	bool bOpenFileDialog,
-	LPCTSTR lpszDefExt,
-	DWORD dwFlags,
-	LPCTSTR lpszFilter,
-	HWND hParentWnd)
-	: m_bOpen(bOpenFileDialog),
-	m_defExt(lpszDefExt ? lpszDefExt : ""),
-	m_filter(lpszFilter ? lpszFilter : ""),
-	m_parent(hParentWnd),
-	m_flags(dwFlags)
-{
-	m_path[0] = 0;
-	m_saveFileName = "";
-}
-
-INT_PTR CFileDialogClsid::DoModal()
-{
-    return DoModalOpen();
-}
-
-void CFileDialogClsid::SetSaveFileName(CString input) {
-    m_saveFileName = input;
-}
-
-CString CFileDialogClsid::GetFilePath() { return m_path; }
-
-CString CFileDialogClsid::GetFileName()
-{
-    CString full = m_path;
-    int pos = full.ReverseFind(L'\\');
-    return pos < 0 ? full : full.Mid(pos + 1);
-}
-
-CString CFileDialogClsid::GetFileExt()
-{
-    CString name = GetFileName();
-    int pos = name.ReverseFind(L'.');
-    return pos < 0 ? CString() : name.Mid(pos + 1);
-}
-
-CString CFileDialogClsid::GetFolderPath()
-{
-    CString full = m_path;
-    int pos = full.ReverseFind(L'\\');
-    return pos < 0 ? CString() : full.Left(pos);
-}
-
-
-INT_PTR CFileDialogClsid::DoModalOpen()
-{
-    CLSID clsid = CLSID_FileOpenDialog;
-    if (!m_bOpen)
-    {
-        clsid = CLSID_FileSaveDialog;
-    }
-    HRESULT hr = CoCreateInstance(clsid, nullptr,
-        CLSCTX_ALL, IID_PPV_ARGS(&m_pDlg));
-    if (FAILED(hr))
-        return IDCANCEL;
-
-	std::vector<std::wstring> storage;
-	if (!m_bOpen) {
-		std::wstring fileName = utf8ToUtf16(m_saveFileName);
-		storage.push_back(fileName);
-		m_pDlg->SetFileName(storage[storage.size() - 1].c_str());
-	}
-	storage.clear();
-
-	std::vector<COMDLG_FILTERSPEC> specs;
-	std::vector<CString> tokens = Split(m_filter, '|');
-
-    // m_filter may have trailing '|'
-	for (size_t i = 0; i < tokens.size(); ++i) {
-		if (tokens[i].IsEmpty()) {
-			tokens.erase(tokens.begin() + i);
-			--i;
-        }
-	}
-	for (size_t i = 0; i + 1 < tokens.size(); i += 2) {
-		auto name = utf8ToUtf16(tokens[i]);
-		auto filter = utf8ToUtf16(tokens[i + 1]);
-		storage.push_back(name);
-		storage.push_back(filter);
-		specs.push_back({
-			storage[storage.size() - 2].c_str(),
-			storage[storage.size() - 1].c_str() 
-		});
-	}
-	m_pDlg->SetFileTypes(static_cast<UINT>(specs.size()), specs.data());
-	storage.clear();
-
-    if (!m_defExt.IsEmpty())
-    {
-        std::wstring defExt = utf8ToUtf16(m_defExt);
-        m_pDlg->SetDefaultExtension(defExt.c_str());
-    }
-
-    DWORD dwOpt = 0;
-    m_pDlg->GetOptions(&dwOpt);
-    m_pDlg->SetOptions(dwOpt | m_flags);
-
-
-    hr = m_pDlg->Show(m_parent);
-    if (hr == HRESULT_FROM_WIN32(ERROR_CANCELLED))
-        return IDCANCEL;
-    if (FAILED(hr))
-        return IDCANCEL;
-
-    CComPtr<IShellItem> pItem;
-    if (FAILED(m_pDlg->GetResult(&pItem)))
-        return IDCANCEL;
-
-    PWSTR psz = nullptr;
-    if (FAILED(pItem->GetDisplayName(SIGDN_FILESYSPATH, &psz)))
-        return IDCANCEL;
-
-    wcscpy_s(m_path, _countof(m_path), psz);
-    CoTaskMemFree(psz);
-    return IDOK;
-}
-
